+++ lua/custom/config/init.lua
@@ -1,6 +1,3 @@
--- 主题的配置必须最优先，否则会覆盖之后的颜色组设置。
-require 'custom.config.onedark'
-
 require 'custom.config.highlight'
 
 require 'custom.config.treesitter'